Replace `rails` dependency with `actioncable` + `railties`
By making `rails` a dependency, it means you can't use this project
in applications using only parts of Rails (e.g. API-only apps, or those
which don't use sprockets, mail, etc). As a gem we should try and
require only the bare minimum required to function.

I believe we only need actioncable and railties for this engine to
function properly. It also looked like activestorage isn't actually used
anywhere so I've also removed that, but we can easily add that as a
development-only dependency if preferred.