Manually prefix route helper in application helper
It looks like this was done in many other places in
c8cf162dd41f42e5d405503764dc007a7a7c97d5.

I'm not _sure_ but I think the difference between our app and the dummy
and demo apps is that we set:

    config.action_controller.include_all_helpers = false

Therefore PageHelper isn't automatically loaded for the
PagesController. The ApplicationHelper is manually included in
ApplicationController, but the PageHelper isn't manually included in the
PagesController, where page_path is being called by the landing_path
method, which is being called when rending the branding partial, which
happens as part of PagesController#show.