fix: allow to nest components in components
Right now component previews cannot be nested within a module that has
the same name as a component (e.g. `Parent::ChildComponentPreview` when
`ParentComponentPreview` already exists.)

This is supported in viewbook itself, but fails in lookbook.

The fix is simple: Within a hierarchical collection any entity could
have content, but it can also have children at the same time.