fix: Handle parameter defaults containing newlines
It seems that AlpineJS doesn't like unescaped newlines in its data
attributes. I was getting JS errors when trying to pass copy with
newlines as the default value of a parameter. This comes up for us
because we have a number of components that handle markdown text. This
fixes the bug (perhaps crudely) and adds a couple tests to cover the
usage cases. The new system test is the only one that was actually
failing with the old implementation, both of the ruby-only tests were
also passing before.