Disable code reloading in production by default
By default a Rails app configs have `cache_classes` as false in
develoment and true in production, and does not change
`reload_classes_only_on_change` from its default `true`.

This means that Lookbook currently actually watches for changes in
Production as well as Development, because it's not taking `cache_classes`
into account.

To quote the Rails Guides:
The Reloader only checks for file changes when `cache_classes` is false and
`reload_classes_only_on_change` is true (which is the default in the `development`
environment).

When `cache_classes` is true (in `production`, by default), the Reloader is
only a pass-through to the Executor.