docs: use Hash#dig in docs ERB
If one uses the suggested layout for the ViewComponent default layout,
navigating to `localhost:3000/rails/view_components` will throw an error
due to the deeply nested nature of the Lookbook params hash. Using `dig`
allows for safe access into the hashes without causing problems if any
of the keys don't exist. This sweeps other instances of deep hash
access.

Hash#dig was introduced in Ruby 2.3, so will be available in all
supported versions of Lookbook.